func: yesterday 함수 정의 및 구현
